more light refactoring
